refactor: separate classes/functions into files
